+++ src/store/index.js
@@ -1,6 +1,7 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 import AuthModule from './modules/AuthModule'
+import TaskModule from './modules/TaskModule'
 Vue.use(Vuex);
 export default new Vuex.Store({
     state: {},
@@ -8,7 +9,8 @@ export default new Vuex.Store({
     mutations: {},
     actions: {},
     modules:{
-        AuthModule
+        AuthModule,
+        TaskModule
     }
 
 });
